requirements.txt
requests
psycopg2
dotenv
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
redis>=4.5.0
celery>=5.3.0
pydantic>=2.0.0
python-decouple
huggingface_hub[hf_xet]
